File check update

+++ run_mpi.sh
@@ -1,7 +1,11 @@
 mpicc -o $1 src/MPI/$1.c
-N=2
+
+N=3
 if [ $# -gt 1 ]; then
    N=$2
 fi
-mpiexec -n $N ./$1
-rm $1
+
+if [ -f $1 ]; then
+   mpiexec -n $N ./$1
+   rm $1
+fi
